+++ bazel_codegen/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.3.4
 
 - Fix the exceptions thrown when attempting to read invalid assets.
+- Return empty sources for unknown assets instead of null.
 
 # 0.3.3+1